requirements.txt
Django==2.0.8
django-tagulous==0.13.2
djangorestframework==3.8.2
mysql-connector-python==8.0.13
mysqlclient==1.3.13
protobuf==3.6.1
pytz==2018.5
six==1.11.0
